Better accuray for diagrams having just a few more values than the diagram's width
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/trunk/kdchart@10558 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ src/KDChartCartesianDiagramDataCompressor_p.h
@@ -91,7 +91,7 @@ namespace KDChart {
         CachePosition mapToCache( const QModelIndex& ) const;
         CachePosition mapToCache( int row, int column ) const;
         QModelIndexList mapToModel( const CachePosition& ) const;
-        int indexesPerPixel() const;
+        qreal indexesPerPixel() const;
 
         // retrieve data from the model, put it into the cache
         void retrieveModelData( const CachePosition& ) const;
